+++ src/AzureAuth/CommandMain.cs
@@ -36,7 +36,6 @@ namespace Microsoft.Authentication.AzureAuth
         private const string OutputOption = "--output";
         private const string AliasOption = "--alias";
         private const string ConfigOption = "--config";
-        private const string CacheOption = "--cache";
         private const string PromptHintPrefix = "AzureAuth";
         private const string TimeoutOption = "--timeout";
 
@@ -68,7 +67,6 @@ Allowed values: [all, web, devicecode]";
         /// The maximum time we will wait to acquire a mutex around prompting the user.
         /// </summary>
         private TimeSpan promptMutexTimeout = TimeSpan.FromMinutes(15);
-        private string cacheFilePath;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="CommandMain"/> class.
@@ -176,40 +174,6 @@ Allowed values: [all, web, devicecode]";
         public double Timeout { get; set; } = GlobalTimeout.TotalMinutes;
 
         /// <summary>
-        /// Gets or sets the cache file name. Only available on Windows.
-        /// </summary>
-        [Option(CacheOption, "Override the default cache file location. This option is only available on Windows.", CommandOptionType.SingleValue, ShowInHelpText = false)]
-        [LegalFilePath]
-        public string CacheFilePath
-        {
-            get
-            {
-                // Check command parameter first.
-                if (!string.IsNullOrEmpty(this.cacheFilePath))
-                {
-                    return this.cacheFilePath;
-                }
-
-                // Check environment variable.
-                string envCacheFile = this.env.Get(EnvVars.Cache);
-                if (!string.IsNullOrEmpty(envCacheFile))
-                {
-                    return envCacheFile;
-                }
-
-                // Use default cache file path.
-                string appData = Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData);
-                string absolutePath = this.fileSystem.Path.Combine(appData, ".IdentityService", $"msal_{this.authSettings.Tenant}.cache");
-                return absolutePath;
-            }
-
-            set
-            {
-                this.cacheFilePath = value;
-            }
-        }
-
-        /// <summary>
         /// Gets the token fetcher options.
         /// </summary>
         public Alias TokenFetcherOptions
@@ -217,7 +181,25 @@ Allowed values: [all, web, devicecode]";
             get { return this.authSettings; }
         }
 
-        private AuthMode CombinedAuthMode => this.AuthModes.Aggregate((a1, a2) => a1 | a2);
+        /// <summary>
+        /// Gets the CombinedAuthMode depending on env variables to disable interactive auth modes.
+        /// </summary>
+        public AuthMode CombinedAuthMode
+        {
+            get
+            {
+                if (this.InteractiveAuthDisabled())
+                {
+#if PlatformWindows
+                    return AuthMode.IWA;
+#else
+                    return 0;
+#endif
+                }
+
+                return this.AuthModes.Aggregate((a1, a2) => a1 | a2);
+            }
+        }
 
         /// <summary>
         /// Combine the <see cref="PromptHintPrefix"/> with the caller provided prompt hint.
@@ -313,6 +295,11 @@ Allowed values: [all, web, devicecode]";
                     Alias configFileOptions = config.Alias[this.AliasName];
                     evaluatedOptions = configFileOptions.Override(evaluatedOptions);
                 }
+                catch (System.IO.FileNotFoundException)
+                {
+                    this.logger.LogError($"The file '{fullConfigPath}' does not exist.");
+                    return false;
+                }
                 catch (Tomlyn.TomlException ex)
                 {
                     this.logger.LogError($"Error parsing TOML in config file at '{fullConfigPath}':\n{ex.Message}");
@@ -346,16 +333,18 @@ Allowed values: [all, web, devicecode]";
             this.eventData.Add("settings_resource", this.authSettings.Resource);
             this.eventData.Add("settings_tenant", this.authSettings.Tenant);
             this.eventData.Add("settings_prompthint", this.authSettings.PromptHint);
-            this.eventData.Add("settings_cachefile", this.CacheFilePath);
 
             // Small bug in Lasso - Add does not accept a null IEnumerable here.
             this.eventData.Add("settings_scopes", this.authSettings.Scopes ?? new List<string>());
 
-            if (this.PCADisabled())
+            if (this.InteractiveAuthDisabled())
             {
-                this.eventData.Add("no_user", true);
-                this.logger.LogCritical($"User based authentication is disabled");
-                return 1;
+                this.eventData.Add(EnvVars.CorextNonInteractive, this.env.Get(EnvVars.CorextNonInteractive));
+                this.eventData.Add(EnvVars.NoUser, this.env.Get(EnvVars.NoUser));
+                this.logger.LogWarning($"Interactive authentication is disabled.");
+#if PlatformWindows
+                this.logger.LogWarning($"Supported auth mode is Integrated Windows Authentication");
+#endif
             }
 
             return this.ClearCache ? this.ClearLocalCache() : this.GetToken();
@@ -365,7 +354,7 @@ Allowed values: [all, web, devicecode]";
         /// Determines whether Public Client Authentication (PCA) is disabled or not.
         /// </summary>
         /// <returns>A boolean to indicate PCA is disabled.</returns>
-        public bool PCADisabled()
+        public bool InteractiveAuthDisabled()
         {
             return !string.IsNullOrEmpty(this.env.Get(EnvVars.NoUser)) ||
                 string.Equals("1", this.env.Get(EnvVars.CorextNonInteractive));
@@ -400,21 +389,13 @@ Allowed values: [all, web, devicecode]";
                 validOptions = false;
             }
 
-            if (!this.CacheFilePath.IsValidAbsoluteFilePath())
-            {
-                this.logger.LogError($"The option {CacheOption}=`{this.CacheFilePath}` " +
-                    $"or environment varable {EnvVars.Cache}=`{this.env.Get(EnvVars.Cache)}` " +
-                    $"is not a valid absolute file path.");
-                validOptions = false;
-            }
-
             return validOptions;
         }
 
         private int ClearLocalCache()
         {
             var pca = PublicClientApplicationBuilder.Create(this.authSettings.Client).Build();
-            var pcaWrapper = new PCAWrapper(this.logger, pca, new List<Exception>(), new Guid(this.authSettings.Tenant), "azureauth", this.CacheFilePath);
+            var pcaWrapper = new PCAWrapper(this.logger, pca, new List<Exception>(), new Guid(this.authSettings.Tenant));
 
             var accounts = pcaWrapper.TryToGetCachedAccountsAsync().Result;
             while (accounts.Any())
@@ -553,10 +534,8 @@ Allowed values: [all, web, devicecode]";
                 new Guid(this.authSettings.Client),
                 new Guid(this.authSettings.Tenant),
                 scopes,
-                this.CacheFilePath,
                 this.PreferredDomain,
-                PrefixedPromptHint(this.authSettings.PromptHint),
-                Constants.AuthOSXKeyChainSuffix);
+                PrefixedPromptHint(this.authSettings.PromptHint));
             }
 
             this.authFlowExecutor = new AuthFlowExecutor(this.logger, authFlows, this.StopwatchTracker());