+++ src/MSALWrapper.Test/AuthFlow/BrokerTest.cs
@@ -39,7 +39,6 @@ namespace Microsoft.Authentication.MSALWrapper.Test
         private Mock<IPCAWrapper> pcaWrapperMock;
         private Mock<IAccount> testAccount;
         private IEnumerable<string> scopes = new string[] { $"{ResourceId}/.default" };
-        private string cacheFilePath = $"Z:/test_cache_file.cache";
         private TokenResult tokenResult;
 
         [SetUp]
@@ -69,7 +68,7 @@ namespace Microsoft.Authentication.MSALWrapper.Test
              .AddTransient<AuthFlow.Broker>((provider) =>
              {
                  var logger = provider.GetService<ILogger<AuthFlow.Broker>>();
-                 return new AuthFlow.Broker(logger, ClientId, TenantId, this.scopes, this.cacheFilePath, pcaWrapper: this.pcaWrapperMock.Object, promptHint: this.promptHint);
+                 return new AuthFlow.Broker(logger, ClientId, TenantId, this.scopes, pcaWrapper: this.pcaWrapperMock.Object, promptHint: this.promptHint);
              })
              .BuildServiceProvider();