+++ src/MSALWrapper.Test/AuthFlow/DeviceCodeTest.cs
@@ -43,7 +43,6 @@ namespace Microsoft.Authentication.MSALWrapper.Test
         private Mock<IAccount> testAccount;
         private IEnumerable<string> scopes = new string[] { $"{ResourceId}/.default" };
         private TokenResult tokenResult;
-        private string cacheFilePath = $"Z:/test_cache_file.cache";
 
         [SetUp]
         public void Setup()
@@ -71,7 +70,7 @@ namespace Microsoft.Authentication.MSALWrapper.Test
              .AddTransient<AuthFlow.DeviceCode>((provider) =>
              {
                  var logger = provider.GetService<ILogger<AuthFlow.DeviceCode>>();
-                 return new AuthFlow.DeviceCode(logger, ClientId, TenantId, this.scopes, this.cacheFilePath, pcaWrapper: this.pcaWrapperMock.Object, promptHint: this.promptHint);
+                 return new AuthFlow.DeviceCode(logger, ClientId, TenantId, this.scopes, pcaWrapper: this.pcaWrapperMock.Object, promptHint: this.promptHint);
              })
              .BuildServiceProvider();