+++ src/MSALWrapper/PCACache.cs
@@ -18,6 +18,7 @@ namespace Microsoft.Authentication.MSALWrapper
         // OSX
         private const string MacOSAccountName = "MSALCache";
         private const string MacOSServiceName = "Microsoft.Developer.IdentityService";
+        private const string OSXKeyChainCategory = "azureauth";
 
         // Linux
         private const string LinuxKeyRingSchema = "com.microsoft.identity.tokencache";
@@ -37,20 +38,13 @@ namespace Microsoft.Authentication.MSALWrapper
         /// </summary>
         /// <param name="logger">The logger.</param>
         /// <param name="tenantId">The tenant id.</param>
-        /// <param name="cacheFilePath">The cache file name.</param>
-        /// <param name="osxKeyChainSuffix">The osx key chain suffix.</param>
-        internal PCACache(ILogger logger, Guid tenantId, string cacheFilePath, string osxKeyChainSuffix = null)
+        internal PCACache(ILogger logger, Guid tenantId)
         {
             this.logger = logger;
-            this.osxKeyChainSuffix = string.IsNullOrWhiteSpace(osxKeyChainSuffix) ? $"{tenantId}" : $"{osxKeyChainSuffix}.{tenantId}";
+            this.osxKeyChainSuffix = $"{OSXKeyChainCategory}.{tenantId}";
 
-            if (string.IsNullOrWhiteSpace(cacheFilePath))
-            {
-                throw new ArgumentNullException($"{nameof(cacheFilePath)} should not be null or whitespace.");
-            }
-
-            this.cacheFileName = Path.GetFileName(cacheFilePath);
-            this.cacheDir = Directory.GetParent(cacheFilePath).FullName;
+            this.cacheFileName = $"msal_{tenantId}.cache";
+            this.cacheDir = this.GetCacheServiceFolder();
         }
 
         /// <summary>
@@ -66,7 +60,7 @@ namespace Microsoft.Authentication.MSALWrapper
                 return;
             }
 
-            var osxKeychainItem = MacOSServiceName + (string.IsNullOrWhiteSpace(this.osxKeyChainSuffix) ? string.Empty : $".{this.osxKeyChainSuffix}");
+            var osxKeychainItem = $"{MacOSServiceName}.{this.osxKeyChainSuffix}";
 
             var storageProperties = new StorageCreationPropertiesBuilder(this.cacheFileName, this.cacheDir)
             .WithLinuxKeyring(LinuxKeyRingSchema, LinuxKeyRingCollection, LinuxKeyRingLabel, linuxKeyRingAttr1, linuxKeyRingAttr2)
@@ -92,6 +86,17 @@ namespace Microsoft.Authentication.MSALWrapper
                 this.logger.LogError("An unexpected error occured creating the cache.");
                 throw new Exception(exceptionMessage);
             }
+        }
+
+        /// <summary>
+        /// Gets the absolute path of the cache folder. Only available on Windows.
+        /// </summary>
+        /// <returns>The absolute path of the cache folder.</returns>
+        private string GetCacheServiceFolder()
+        {
+            string appData = Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData);
+            string absolutePath = Path.Combine(appData, ".IdentityService");
+            return absolutePath;
         }
     }
 }